requirements.txt
streamlit==1.56.0
fastapi==0.135.1
sqlalchemy==2.0.48
biopython==1.86
requests==2.32.5
python-docx==1.1.2
